clientes/requirements.txt
kafka-python
psycopg2-binary
tweepy